Split relay policy boundaries
